refactor: cardHideOption.types.ts -> cardHideOptionSelect feature 이하로 이동
